make task runtime human-readable
