+++ src/Illuminate/Console/View/Components/Task.php
@@ -2,6 +2,7 @@
 
 namespace Illuminate\Console\View\Components;
 
+use Illuminate\Support\InteractsWithTime;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
 
@@ -9,6 +10,8 @@ use function Termwind\terminal;
 
 class Task extends Component
 {
+    use InteractsWithTime;
+
     /**
      * Renders the component using the given arguments.
      *
@@ -39,7 +42,7 @@ class Task extends Component
             throw $e;
         } finally {
             $runTime = $task
-                ? (' '.number_format((microtime(true) - $startTime) * 1000).'ms')
+                ? (' '.$this->formatRunTime($startTime))
                 : '';
 
             $runTimeWidth = mb_strlen($runTime);
