+++ src/Illuminate/Support/InteractsWithTime.php
@@ -2,6 +2,7 @@
 
 namespace Illuminate\Support;
 
+use Carbon\CarbonInterval;
 use DateInterval;
 use DateTimeInterface;
 
@@ -60,5 +61,20 @@ trait InteractsWithTime
     protected function currentTime()
     {
         return Carbon::now()->getTimestamp();
+    }
+
+    /**
+     * Given a start time, format the total run time for human readability.
+     *
+     * @param  float  $startTime
+     * @return string
+     */
+    protected function formatRunTime($startTime)
+    {
+        $runTime = (microtime(true) - $startTime) * 1000;
+
+        return $runTime > 1000
+            ? CarbonInterval::milliseconds($runTime)->cascade()->forHumans(short: true)
+            : number_format($runTime, 2).'ms';
     }
 }
